+++ tests/structinit/structinit.c
@@ -2,6 +2,7 @@
 #include <git2/sys/config.h>
 #include <git2/sys/odb_backend.h>
 #include <git2/sys/refdb_backend.h>
+#include <git2/sys/transport.h>
 
 #define STRINGIFY(s) #s
